+++ trunk/tests/test-fscal.C
@@ -237,11 +237,11 @@ int main(int ac, char **av) {
 			pass &= test_fscal(F,m,k,n,timing);
 		}
 		{
-			Givaro::Modular<int32_t> F(p) ;
+			Givaro::Modular<int32_t> F((int32_t)p) ;
 			pass &= test_fscal(F,m,k,n,timing);
 		}
 		{
-			Givaro::ModularBalanced<int32_t> F((int)p) ;
+			Givaro::ModularBalanced<int32_t> F((int32_t)p) ;
 			pass &= test_fscal(F,m,k,n,timing);
 		}
 		{
@@ -289,11 +289,11 @@ int main(int ac, char **av) {
 			pass &= test_fscalin(F,m,k,n,timing);
 		}
 		{
-			Givaro::Modular<int32_t> F(p) ;
+			Givaro::Modular<int32_t> F((int32_t)p) ;
 			pass &= test_fscalin(F,m,k,n,timing);
 		}
 		{
-			Givaro::ModularBalanced<int32_t> F((int)p) ;
+			Givaro::ModularBalanced<int32_t> F((int32_t)p) ;
 			pass &= test_fscalin(F,m,k,n,timing);
 		}
 		{
